Refactor the class structure of things so that the mod properly NO-OPs when put on a server instead of crashing
